services/market-data-realtime/requirements.txt
fyers-apiv3
fastapi
uvicorn[standard]
python-dotenv
motor
